parsers.c: avoid implicit conversion loses integer precision warning
This warning is raised by Apple LLVM version 6.0 (clang-600.0.57)
(based on LLVM 3.5svn) and was introduced in 521875cc28be

+++ mercurial/parsers.c
@@ -189,7 +189,7 @@ static PyObject *make_file_foldmap(PyObject *self, PyObject *args)
 			      &PyFunction_Type, &normcase_fallback))
 		goto quit;
 
-	spec = PyInt_AS_LONG(spec_obj);
+	spec = (int)PyInt_AS_LONG(spec_obj);
 	switch (spec) {
 	case NORMCASE_LOWER:
 		table = lowertable;
